refactor: Responsiveness and icon adjustment
