+++ lib/app_widget.dart
@@ -1,5 +1,5 @@
+import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:projeto_criptos/src/presenter/screens/wallet/wallet_page.dart';
 
 class MyApp extends StatelessWidget {
@@ -7,15 +7,12 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ProviderScope(
-      child: MaterialApp(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          primarySwatch: Colors.blue,
-        ),
-        home: const WalletScreen(),
-        debugShowCheckedModeBanner: false,
-      ),
+    return MaterialApp(
+      useInheritedMediaQuery: true,
+      locale: DevicePreview.locale(context),
+      builder: DevicePreview.appBuilder,
+      home: const WalletScreen(),
+      debugShowCheckedModeBanner: false,
     );
   }
 }
