+++ lib/main.dart
@@ -1,7 +1,14 @@
+import 'package:device_preview/device_preview.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import 'app_widget.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(DevicePreview(
+      enabled: !kReleaseMode,
+      builder: (context) {
+        return const ProviderScope(child: MyApp());
+      }));
 }
