+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   hooks_riverpod: ^1.0.4
   intl: ^0.17.0
+  device_preview: ^1.1.0
 
 dev_dependencies:
   flutter_test:
